# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: DishFlow API
  version: 1.0.0
  description: 多门店餐饮点餐与履约系统 HTTP 契约（不含 iOS）
servers:
  - url: /
...
paths:
...
  /media/menu/{path}:
    get:
      tags: [customer]
      summary: 可信菜单图片代理
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: image
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/auth/wechat/session:
    post:
      tags: [customer]
      summary: 微信 code 换顾客会话
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: session
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  customer_id: { type: string }
                  store_id: { type: string }
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/storefront/bootstrap:
    get:
      tags: [customer]
      summary: 门店启动信息
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      responses:
        "200":
          description: bootstrap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bootstrap"
  /api/v1/store:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      responses:
        "200":
          description: store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicStore"
  /api/v1/store/policies:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      responses:
        "200":
          description: policies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policies"
  /api/v1/menu:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      responses:
        "200":
          description: menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerMenu"
  /api/v1/tables/resolve:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - in: query
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: table
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/pickup-slots:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - in: query
          name: date
          schema: { type: string, format: date }
      responses:
        "200":
          description: slots
  /api/v1/coupon-offers:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      responses:
        "200":
          description: offers
  /api/v1/pricing/preview:
    post:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreviewRequest"
      responses:
        "200":
          description: quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewResponse"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/orders:
    post:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quote_token]
              properties:
                quote_token: { type: string }
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/CartItem"
                remark: { type: string, maxLength: 100 }
      responses:
        "200":
          description: order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: orders
  /api/v1/orders/{id}:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /api/v1/orders/{id}/prepay:
    post:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: prepay
  /api/v1/orders/{id}/mock-payment/confirm:
    post:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: confirmed
  /api/v1/orders/{id}/cancel:
    post:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: cancel
  /api/v1/coupons:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: coupons
  /api/v1/coupon-offers/{template_id}/claim:
    post:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - in: path
          name: template_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: claimed
  /api/v1/me:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      responses:
        "200":
          description: me
  /api/v1/me/membership:
    post:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone_code: { type: string }
                agreed: { type: boolean }
      responses:
        "200":
          description: member
  /api/v1/me/points:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      responses:
        "200":
          description: ledger
  /api/v1/me/rewards:
    get:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
      responses:
        "200":
          description: rewards
  /api/v1/me/rewards/{template_id}/redeem:
    post:
      tags: [customer]
      parameters:
        - $ref: "#/components/parameters/WechatAppid"
        - in: path
          name: template_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: redeemed
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/admin/session:
    post:
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login_name, password]
              properties:
                login_name: { type: string }
                password: { type: string }
      responses:
        "200":
          description: principal
    get:
      tags: [admin]
      responses:
        "200":
          description: principal
    delete:
      tags: [admin]
      responses:
        "200":
          description: logged out
...
  /api/v1/admin/platform/stores/{id}:
    patch:
      tags: [platform]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: updated }
...
  /api/v1/admin/platform/users/{id}:
    patch:
      tags: [platform]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: updated }
  /api/v1/admin/platform/users/{id}/assign-store-owner:
    post:
      tags: [platform]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: assigned }
...
  /api/v1/admin/platform/shop-applications/{id}/review:
    post:
      tags: [platform]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: reviewed }
  /api/v1/admin/orders/board:
    get:
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/StoreId"
      responses:
        "200": { description: board }
  /api/v1/admin/orders:
    get:
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/StoreId"
      responses:
        "200": { description: list }
  /api/v1/admin/orders/export:
    get:
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/StoreId"
      responses:
        "200":
          description: csv
          content:
            text/csv:
              schema: { type: string }
  /api/v1/admin/orders/{id}:
    get:
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/StoreId"
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: order }
  /api/v1/admin/orders/{id}/transitions:
    post:
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/StoreId"
        - $ref: "#/components/parameters/ID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                to_status: { type: string }
                expected_version: { type: integer }
      responses:
        "200": { description: order }
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/admin/orders/{id}/refunds:
    post:
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/StoreId"
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { description: refund }
...
  /api/v1/admin/refunds:
    get:
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/StoreId"
      responses:
        "200": { description: list }
...
components:
  parameters:
    WechatAppid:
      in: header
      name: X-Wechat-Appid
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Bootstrap:
      type: object
      properties:
        store_id: { type: string }
        store_name: { type: string }
        brand_name: { type: string }
        theme_color: { type: string }
        logo_url: { type: string }
        is_open: { type: boolean }
        announcement: { type: string }
    PublicStore:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        address: { type: string }
        business_hours: { type: string }
        is_open: { type: boolean }
        timezone: { type: string }
    Policies:
      type: object
      properties:
        privacy_policy: { type: string }
        refund_policy: { type: string }
        qualifications: { type: string }
...
    PreviewRequest:
      type: object
      required: [scene, items]
      properties:
        scene: { type: string, enum: [DINE_IN, PICKUP] }
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        table_token: { type: string }
        scheduled_for: { type: string, format: date-time }
        customer_coupon_id: { type: string }
...